TLRC - 999 - Implement Error Reporting for Failed Validations (Stage 4 Validations)  - TLRC - 1067 - Dev - Rename Uln Datatype to Long

+++ src/Sfa.Tl.ResultsAndCertification.Models/Registration/RegistrationRecordResponse.cs
@@ -12,7 +12,7 @@ namespace Sfa.Tl.ResultsAndCertification.Models.Registration
             TlSpecialismLarIds = new List<KeyValuePair<int, string>>();
         }
 
-        public int Uln { get; set; }
+        public long Uln { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime DateOfBirth { get; set; }
